% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataAnalysis.R
\name{analyse_biodiversity}
\alias{analyse_biodiversity}
\title{Perform diversity calculations}
\usage{
analyse_biodiversity(
  Hyperspectral_Image_File_Path,
  Mask_Image_File_Path,
  NBbclusters = 20,
  Window_size = 20,
  NbCPU = 4,
  MaxRAM = 8
)
}
\arguments{
\item{Hyperspectral_Image_File_Path}{character. Path of the image to be processed}

\item{Mask_Image_File_Path}{character. Path of the mask file image}
}
\value{
Returns the full rectified image file path
}
\description{
This function is a wrapper for the biodiversity calculations using the
package BiodivmapR.
}
