% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRectangleFromPlotlocations.R
\name{create_rectangular_shapefile}
\alias{create_rectangular_shapefile}
\title{Create Rectangular shapefile from plot locations}
\usage{
create_rectangular_shapefile(
  Plotlocation_Shp_File_Path,
  Output_Folder_Path,
  Buffer = 10,
  Square = FALSE
)
}
\arguments{
\item{Plotlocation_Shp_File_Path}{character. Path of the plot locations shp}

\item{Output_Folder_Path}{character. Folder path of the shp of the cut file image}

\item{Buffer}{character. How much buffer is to be set around the plot locations}
}
\value{
Returns the full file path of the cut file
}
\description{
This function is used to create a rectangular box around plot locations
}
