% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateRGB.R
\name{create_RGB}
\alias{create_RGB}
\title{Create RGB image}
\usage{
create_RGB(
  Raw_Image_Folder_Path,
  RGB_Image_Folder_Path,
  Site_Name = "Site",
  Band_Selection = "-b 59 -b 34 -b 20",
  Output_Image_Format = "GTiff"
)
}
\arguments{
\item{Raw_Image_Folder_Path}{character. Path of the image to be processed}

\item{RGB_Image_Folder_Path}{character. Path of the RGB image}

\item{Site_Name}{character. Name of the test site}

\item{Band_Selection}{character. Bandselection in the format "-b 59 -b 34 -b 20"}

\item{Output_Image_Format}{Format ot the RGB image (default: GTiff)}
}
\value{
Returns the full hyperspectral image path
}
\description{
This function is used to create a RGB image out of a hyperspectral image
}
