% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSAVIMask.R
\name{create_SAVI_mask}
\alias{create_SAVI_mask}
\title{Create SAVI mask}
\usage{
create_SAVI_mask(
  Hyperspectral_Image_File_Path,
  SAVI_Mask_Folder_Path,
  SAVI_threshold = 0.2,
  SAVI_L = 0.5
)
}
\arguments{
\item{Hyperspectral_Image_File_Path}{character. Path of the image to be processed}

\item{SAVI_Mask_Folder_Path}{character. Path of the rectified image}

\item{SAVI_threshold}{number. SAVI threshold. Default: 0.2}

\item{SAVI_L}{number. Canopy background adjustment factor. Default: 0.5}
}
\value{
Returns the full mask file path
}
\description{
This Library is used to create a SAVI (soil adjusted vegetation index)
mask for a hyperspectral image
}
