% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RectifyImage.R
\name{rectify_Image}
\alias{rectify_Image}
\title{Rectify hyperspectral image}
\usage{
rectify_Image(
  Hyperspectral_Raw_Image_Folder_Path,
  Rectified_Image_Folder_Path,
  Cutfile = FALSE,
  Cutfile_Folder_Path
)
}
\arguments{
\item{Hyperspectral_Raw_Image_Folder_Path}{character. Path of the image to be processed}

\item{Rectified_Image_Folder_Path}{character. Path of the rectified image}

\item{Cutfile}{boolean. If true, the image is cut according to the shp file}

\item{Cutfile_Folder_Path}{character. Path to the shp file}
}
\value{
Returns the full rectified image file path
}
\description{
This function is used to rectify hyperspectral images using gdalwarp
}
