% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClusterTest.R
\name{test_cluster}
\alias{test_cluster}
\title{Perform cluster testing for kmeans clustering}
\usage{
test_cluster(Main_Folder_Path, NbCPU = 4, MaxRAM = 8, PCA_Threshold = 99)
}
\arguments{
\item{Main_Folder_Path}{character. Path of the main folder}

\item{NbCPU}{numeric. Number of CPUs to use in parallel.}

\item{MaxRAM}{numeric. MaxRAM maximum size of chunk in GB to limit RAM allocation when reading image file.}

\item{PCA_Threshold}{number. Percentage explained by PCs for selecting PCs.}
}
\value{
Returns the full file name
}
\description{
This function is a wrapper for the biodiversity calculations using the
package BiodivmapR.
}
