% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WssClusterAnalysis.R
\name{get_optimal_cluster_number}
\alias{get_optimal_cluster_number}
\title{Perform cluster analysis with wss}
\usage{
get_optimal_cluster_number(Image_File_Path, Min_Cluster = 2, Max_Cluster = 30)
}
\arguments{
\item{Image_File_Path}{character. Path of the selected PC image}

\item{Min_Cluster}{numeric. Minimal amount of clusters to test.}

\item{Max_Cluster}{numeric. Maximal amount of clusters to test.}
}
\value{
Returns the number of optimal clusters
}
\description{
This uses wss to get the optimal kmeans cluster number.
}
